Refactor to add a Diff class
