+++ lib/Metadata.js
@@ -1,12 +1,5 @@
 const path = require('path')
 const conceptStore = require('../lib/store/concept-store')
-const nunjucks = require('nunjucks')
-const jsondiffpatch = require('jsondiffpatch').create({
-    textDiff: {
-        minLength: 60
-    }
-})
-const formatters = require('jsondiffpatch').formatters
 
 class Metadata {
     constructor(concepts, descriptors, messages) {
@@ -83,32 +76,6 @@ class Metadata {
             }
         }
         return false
-    }
-
-    diffReady() {
-        const ready = {}
-        for (const conceptId of Object.keys(this.getConcepts())) {
-            const conceptName = conceptStore.getConceptNameById(conceptId)
-            const newKey = `${conceptName} (${conceptId})`
-            ready[newKey] = this.getConcept(conceptId)
-        }
-        return ready
-    }
-
-    diff(metadata, title, date) {
-        // Perform a diff using self as "a" and the passed argument as "b".
-        const delta = jsondiffpatch.diff(this.diffReady(), metadata.diffReady())
-        const diff = formatters.html.format(delta)
-        const layoutFolder = path.join(__dirname, 'layout')
-        const layout = 'diff.njk'
-        const nunjucksOptions = { autoescape: false }
-        const layoutData = {
-            title: title,
-            date: date,
-            diff: diff
-        }
-        nunjucks.configure(layoutFolder, nunjucksOptions)
-        return nunjucks.render(layout, layoutData)
     }
 }
