+++ lib/index.js
@@ -12,6 +12,7 @@ const descriptorStore = require('./store/descriptor-store')
 // Utility classes
 const Metadata = require('./Metadata')
 const InputError = require('./InputError')
+const Diff = require('./Diff')
 
 module.exports = {
     WordTemplateInput,
@@ -24,4 +25,5 @@ module.exports = {
     descriptorStore,
     Metadata,
     InputError,
+    Diff,
 }
