+++ lib/layout/diff.njk
@@ -32,6 +32,6 @@
     <body>
         <h1>{{ title }}</h1>
         <h3>{{ date }}</h3>
-        {{ diff }}
+        {{ htmlDiff }}
     </body>
 </html>
